requirements.txt
torch
torchvision
Pillow
huggingface-hub
numpy
